Move getExamples api to separate examplesController with GET url /api/examples
