+++ server/src/server.js
@@ -5,7 +5,8 @@ const mongoose = require('mongoose');
 mongoose.Promise = Promise;
 const bodyParser = require('body-parser');
 import express from 'express';
-import { userRoutes } from './controllers/userController.js';
+import userRoutes from './controllers/userController';
+import examplesRoutes from './controllers/examplesController';
 const passport = require('passport');
 const cors = require('cors');
 const app = express();
@@ -53,6 +54,7 @@ router.use('/auth', authRoutes);
 router.use('/users', userRoutes);
 router.use('/pages', pageRoutes);
 router.use('/folders', folderRoutes);
+router.use('/examples', examplesRoutes);
 router.use('/', apiRoutes);
 app.use('/api', router);
